mini_program/dynamic_grid_generator/requirements.txt
pillow
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
requests==2.33.0
urllib3==2.6.3
